+++ ormar/models/descriptors/__init__.py
@@ -1,0 +1,17 @@
+from ormar.models.descriptors.descriptors import (
+    BytesDescriptor,
+    JsonDescriptor,
+    PkDescriptor,
+    PropertyDescriptor,
+    PydanticDescriptor,
+    RelationDescriptor,
+)
+
+__all__ = [
+    "PydanticDescriptor",
+    "RelationDescriptor",
+    "PropertyDescriptor",
+    "PkDescriptor",
+    "JsonDescriptor",
+    "BytesDescriptor",
+]